move presets directory to reactjs
