+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/__tests__/replaceElementsWithReferencesInProps.test.ts
@@ -1,4 +1,4 @@
-import { ComponentPresetElementProps } from '../ComponentPreset';
+import { ComponentPresetElementProps } from '../../../../../common/types/ComponentPreset';
 import { PartialProps } from '../jsx/JSXSerializationResult';
 import { replaceElementsWithReferencesInProps } from '../replaceElementsWithReferencesInProps';
