+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/getPresetElementReference.ts
@@ -1,4 +1,4 @@
-import { PresetElementReference } from './ComponentPreset';
+import { PresetElementReference } from '../../../../common/types/ComponentPreset';
 import { isJSXSerializedElement } from './isJSXSerializedElement';
 import { AnySerializedElement } from './jsx/JSXSerializationResult';
