+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/jsx/bundle/__tests__/getSourceFileContentToBundle.test.ts
@@ -1,5 +1,5 @@
 import { FrameworkNames } from '../../../../../../../framework/frameworkNames';
-import { ComponentDefinition } from '../../../../ComponentDefinition';
+import { ComponentDefinition } from '../../../../../../../steps/serialization/component/ComponentDefinition';
 import { getSourceFileContentToBundle } from '../getSourceFileContentToBundle';
 
 describe('getSourceFileContentToBundle', () => {
