+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/jsx/compile/__tests__/getNamespacedComponentsTree.test.ts
@@ -1,5 +1,5 @@
 import { FrameworkNames } from '../../../../../../../framework/frameworkNames';
-import { ComponentDefinition } from '../../../../ComponentDefinition';
+import { ComponentDefinition } from '../../../../../../../steps/serialization/component/ComponentDefinition';
 import { getNamespacedComponentsTree, NamespacedComponentsTree } from '../getNamespacedComponentsTree';
 
 describe('getNamespacedComponentsTree', () => {
