+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/jsx/compile/compilePresets.ts
@@ -4,9 +4,9 @@ import * as webpack from 'webpack';
 import { ProgramArgs, RawProgramArgs } from '../../../../../../program/args/ProgramArgs';
 import { getProjectRoot } from '../../../../../../program/args/providers/paths/getProjectRoot';
 import { getTempDirPath } from '../../../../../../program/args/providers/paths/getTempDirPath';
-import { Compiler } from '../../../../../building/compiler/Compiler';
-import { WebpackCompiler } from '../../../../../building/compiler/webpack/WebpackCompiler';
-import { ComponentDefinition } from '../../../ComponentDefinition';
+import { Compiler } from '../../../../../../steps/building/compiler/Compiler';
+import { WebpackCompiler } from '../../../../../../steps/building/compiler/webpack/WebpackCompiler';
+import { ComponentDefinition } from '../../../../../../steps/serialization/component/ComponentDefinition';
 import { createBundleSource } from '../bundle/createBundleSource';
 import { generateVirtualModules, VirtualComponentModule } from './generateVirtualModules';
 import { getPresetsBundleWebpackConfig } from './getPresetsBundleWebpackConfig';
