+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/jsx/compile/getNamespacedComponentsTree.ts
@@ -1,6 +1,6 @@
 import { set } from 'lodash';
-import { ComponentDefinition } from '../../../ComponentDefinition';
-import { getComponentNamespacedName } from '../../../name/getComponentNamespacedName';
+import { ComponentDefinition } from '../../../../../../steps/serialization/component/ComponentDefinition';
+import { getComponentNamespacedName } from '../../../../../../steps/serialization/component/name/getComponentNamespacedName';
 import { ComponentPlaceholder } from './generateVirtualModules';
 
 export interface NamespacedComponentsTree {
