+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/presets/jsx/compile/globals/__uxpinParsePreset.ts
